Completed save recipe use case
Main Changes:
There is an "Add to Favourites" button when you search recipes and clicking it saves it in a text file.
The saved recipes are displayed in the favourites page

Additional Changes:

Fixed bug where home page didn't open first.
Made it possible to change pages (home, favourite, explore)